consolidated knight class
