+++ Makefile
@@ -1,8 +1,9 @@
 CC=gcc
 CXX=g++
-CXXFLAGS=-g -Wall
+CXXFLAGS=-g -Wall -std=c++11 -stdlib=libc++
 CFLAGS=-g -Wall
 LDFLAGS=-g -Wall
+
 
 .PHONY: build all test run clean
 build: src/main test/codalot_test
@@ -12,7 +13,7 @@ all: clean build
 test: test/codalot_test
 	test/codalot_test
 
-run: src/main
+run: src/main 
 	src/main
 
 clean:
@@ -21,5 +22,8 @@ clean:
 	rm -rf *.o a.out a.exe main main.exe codalot_test codalot_test.exe
 
 test/codalot_test:
+	$(CXX) $(CXXFLAGS) test/codalot_test.cpp src/Knight.cpp -o test/main
 
-src/main:
+src/main: 
+	$(CXX) $(CXXFLAGS) src/main.cpp src/Knight.cpp -o src/main
+
