+++ test/basic_codalot.cpp
@@ -1,74 +1,9 @@
 #include <vector>
 #include <cstdlib>
 #include "codalot.cpp"
+#include "../src/Knight.h"
 
 using namespace std;
-
-class Knight {
-private:
-    int xp;
-    int stamina;
-    bool inTavern;
-    bool inTrainingYard;
-public:
-    Knight();
-
-    int getXp();
-    void setXp(int xp);
-    void incrementXp(int xp);
-    int getStamina();
-    void setStamina(int stamina);
-    void incrementStamina(int stamina);
-    bool isInTavern();
-    void setInTavern(bool inTavern);
-    bool isInTrainingYard();
-    void setInTrainingYard(bool inTrainingYard);
-};
-
-Knight::Knight() {
-    xp = 0;
-    stamina = 0;
-}
-
-int Knight::getXp() {
-    return xp;
-}
-
-void Knight::setXp(int xp) {
-    this->xp = xp;
-}
-
-void Knight::incrementXp(int xp) {
-    this->xp += xp;
-}
-
-int Knight::getStamina() {
-    return stamina;
-}
-
-void Knight::setStamina(int stamina) {
-    this->stamina = stamina;
-}
-
-void Knight::incrementStamina(int stamina) {
-    this->stamina += stamina;
-}
-
-bool Knight::isInTavern() {
-    return inTavern;
-}
-
-void Knight::setInTavern(bool inTavern) {
-    this->inTavern = inTavern;
-}
-
-bool Knight::isInTrainingYard() {
-    return inTrainingYard;
-}
-
-void Knight::setInTrainingYard(bool inTrainingYard) {
-    this->inTrainingYard = inTrainingYard;
-}
 
 class BasicCodalot : public Codalot {
 private:
