+++ test/knight_position.h
@@ -1,0 +1,4 @@
+typedef enum KnightPosition_t {
+    TAVERN,
+    TRAINING_YARD
+} KnightPosition;
